Rename README.md to README2.md
Comfort Boamah
UEB3212022
IT D

 Login and Registration System.
The login And Registration System project in C++ involves mainly the user registration process. User Credentials like usernames and passwords are asked from the user. If the registration of the user is successful then with the given credentials a file will be created of a particular user in the database.
Technologies Required: C++ programming language, User Input/Output Libraries, File Manipulation Libraries, Strings and Structures Libraries and Database Libraries.